+++ api/src/shared/infrastructure/repositories/index.js
@@ -1,8 +1,8 @@
 import * as certificationEvaluationApi from '../../../certification/evaluation/application/api/select-next-certification-challenge-api.js';
+import * as certificationEvaluationRepository from '../../../certification/shared/infrastructure/repositories/certification-challenge-repository.js';
 import * as campaignsAPI from '../../../prescription/campaign/application/api/campaigns-api.js';
 import * as knowledgeElementSnapshotAPI from '../../../prescription/campaign/application/api/knowledge-element-snapshots-api.js';
 import { injectDependencies } from '../utils/dependency-injection.js';
-import * as certificationEvaluationRepository from './certification-challenge-repository.js';
 import * as knowledgeElementRepository from './knowledge-element-repository.js';
 
 /**
